refactor(model/linear): refactor linear module
