+++ internlm/core/parallel/comm/isp.py
@@ -1,18 +1,22 @@
 #!/usr/bin/env python
 # -*- encoding: utf-8 -*-
+"""
+communication for isp parallel.
+"""
 
 from dataclasses import dataclass
 from functools import partial
-from typing import Any, Callable, Dict, List, Union
+from typing import Any, Callable, Dict, List, Tuple, Union
 
 import torch
 from torch import distributed as dist
 from torch import nn
 
 from internlm.core.context import global_context as gpc
+from internlm.core.parallel.comm.utils import DUMMY_HANDLE_CONST, AsyncCommHandle
 from internlm.core.naive_amp import NaiveAMPModel
-from internlm.model.linear import ISPLinear
-from internlm.model.utils import all_gather_raw, reduce_scatter_raw
+from internlm.model.modules.ffn import ISPLinear
+from internlm.model.modules.utils import all_gather_raw, reduce_scatter_raw
 from internlm.utils.common import SchedulerHook
 
 
@@ -470,7 +474,7 @@ class ISPCommunicator:
 
     # communication operation interfaces
 
-    def all_gather(self, tensor: torch.Tensor, module: nn.Module, is_bias: bool = False):
+    def all_gather(self, tensor: torch.Tensor, module: nn.Module, is_bias: bool = False) -> torch.Tensor:
         if dist.get_world_size(self.process_group) <= 1:
             return tensor
 
@@ -489,9 +493,9 @@ class ISPCommunicator:
         model: nn.Module,
         op: dist.ReduceOp,
         is_bias: bool = False,
-    ):
+    )-> Tuple[torch.Tensor, AsyncCommHandle]:
         if dist.get_world_size(self.process_group) <= 1:
-            return tensor, None
+            return tensor, DUMMY_HANDLE_CONST
 
         if not self.overlap:
             result, handle = reduce_scatter_raw(tensor, self.process_group, op=op, async_op=True)
@@ -520,7 +524,7 @@ class ISPCommunicator:
                         *tensor.shape[1:],
                     )
                 ),
-                None,
+                DUMMY_HANDLE_CONST,
             )
 
         return result, handle
