+++ internlm/model/__init__.py
@@ -2,7 +2,7 @@
 # -*- encoding: utf-8 -*-
 
 from .embedding import Embedding1D, RotaryEmbedding
-from .linear import FeedForward, RewardModelLinear, ScaleColumnParallelLinear
+from .modules.ffn import FeedForward, RewardModelLinear, ScaleColumnParallelLinear
 from .metrics import AccPerplex
 from .modeling_internlm import build_model_with_cfg
 from .modeling_internlm2 import build_model_with_cfg as build_model_with_cfg2
@@ -10,7 +10,7 @@ from .modeling_llama import build_model_with_cfg as build_model_with_llama_cfg
 from .modeling_moe import build_model_with_moe_cfg
 from .moe import MoE
 from .multi_head_attention import MHA, DistributedAttention
-from .utils import gather_forward_split_backward
+from .modules.utils import gather_forward_split_backward
 
 __all__ = [
     "Embedding1D",
