+++ internlm/model/modeling_internlm2.py
@@ -20,7 +20,7 @@ from internlm.model.embedding import (
     Embedding1D,
     RotaryEmbedding,
 )
-from internlm.model.linear import (
+from internlm.model.modules.ffn import (
     InternLM2ScaleColumnParallelLinear,
     RewardModelLinear,
     get_linear_cls,
@@ -32,7 +32,7 @@ from internlm.model.multi_head_attention import (
     SelfAttention,
     _update_kv_cache,
 )
-from internlm.model.utils import (
+from internlm.model.modules.utils import (
     gather_forward_split_backward,
     split_forward_gather_backward,
     try_import_RMSNorm,
