+++ internlm/model/moe.py
@@ -3,7 +3,7 @@ import torch
 import internlm.moe  # noqa # pylint: disable=W0611
 from internlm.core.context import ParallelMode
 from internlm.core.context import global_context as gpc
-from internlm.model.linear import FeedForward
+from internlm.model.modules.ffn import FeedForward
 from internlm.utils.logger import get_logger
 from internlm.utils.registry import MODEL_INITIALIZER
