+++ internlm/model/multi_head_attention.py
@@ -14,7 +14,7 @@ from torch.nn import Module
 
 from internlm.core.context import global_context as gpc
 from internlm.model.embedding import DynamicNTKScalingRotaryEmbedding, RotaryEmbedding
-from internlm.model.linear import get_linear_cls
+from internlm.model.modules.ffn import get_linear_cls
 
 
 # adpated from https://github.com/microsoft/DeepSpeed/blob/master/deepspeed/sequence/layer.py
