+++ internlm/train/__init__.py
@@ -2,7 +2,7 @@ from .training_internlm import (
     get_scheduler_hooks,
     get_train_data_loader,
     get_validation_data_loader,
-    initialize_isp_communicator,
+    initialize_parallel_communicator,
     initialize_llm_profile,
     initialize_model,
     initialize_optimizer,
@@ -18,7 +18,7 @@ __all__ = [
     "get_validation_data_loader",
     "initialize_llm_profile",
     "initialize_model",
-    "initialize_isp_communicator",
+    "initialize_parallel_communicator",
     "initialize_optimizer",
     "load_new_batch",
     "record_current_batch_training_metrics",
