+++ tests/test_model/test_model_internlm.py
@@ -10,9 +10,9 @@ import internlm
 from internlm.core.context import ParallelMode
 from internlm.core.context.parallel_context import Config
 from internlm.core.context.parallel_context import global_context as gpc
-from internlm.model.linear import RewardModelLinear, ScaleColumnParallelLinear
+from internlm.model.modules.ffn import RewardModelLinear, ScaleColumnParallelLinear
 from internlm.model.modeling_internlm import PackedFlashBaseLayer1D
-from internlm.model.utils import gather_forward_split_backward
+from internlm.model.modules.utils import gather_forward_split_backward
 
 config = Config(
     dict(
