+++ tests/test_training/test_norm_weight.py
@@ -13,7 +13,7 @@ from internlm.model.metrics import AccPerplex
 from internlm.train import (
     get_scheduler_hooks,
     get_train_data_loader,
-    initialize_isp_communicator,
+    initialize_parallel_communicator,
     initialize_model,
     initialize_optimizer,
 )
@@ -71,7 +71,7 @@ def train_check_norm_weight(args):
     model = initialize_model()
 
     # initialize isp communicator
-    isp_communicator = initialize_isp_communicator(model)
+    isp_communicator = initialize_parallel_communicator(model)
 
     # initialize loss function
     criterion = FlashGPTLMLoss(parallel_output=True, label_smoothing=gpc.config.loss.label_smoothing)
