+++ train.py
@@ -22,7 +22,7 @@ from internlm.train import (
     get_scheduler_hooks,
     get_train_data_loader,
     get_validation_data_loader,
-    initialize_isp_communicator,
+    initialize_parallel_communicator,
     initialize_llm_profile,
     initialize_model,
     initialize_optimizer,
@@ -103,7 +103,7 @@ def main(args):
     model = initialize_model()
 
     # initialize isp communicator
-    isp_communicator = initialize_isp_communicator(model)
+    isp_communicator = initialize_parallel_communicator(model)
 
     with open(args.config, "r") as f:
         config_lines = f.readlines()
